chore: move tauri command invoke arg types into seperate file
